The FIONREAD ioctl() needs an int argument pointer rather than a long argument pointer, since it returns a 32-bit number on both 32-bit and 64-bit platforms.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3180 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_socket.c
@@ -2358,7 +2358,7 @@ mx_socket_num_input_bytes_available( MX_SOCKET *mx_socket,
 	int num_fds, select_status, socket_errno;
 	struct timeval timeout;
 	char *error_string;
-	long num_socket_bytes_available;
+	int num_socket_bytes_available;
 
 	unsigned long flags;
 	mx_status_type mx_status;
